+++ CHANGELOG.md
@@ -55,3 +55,7 @@
    - Fix bad import in `setup.py`
    - Remove requirement for PyTables to enable support on python >3.9 #5
    - Add further save- and load functions for .csv
+- v0.3.0:
+   - Add parallelization option for optimization backends
+- v0.3.1:
+  - Avoid DType Warning #53